+++ CSOPESY-OS-Emulator/src/FlatAllocator.h
@@ -14,7 +14,7 @@ public:
 	size_t allocateAt(std::shared_ptr<Process> processAddress, size_t index);
 	void deallocateAt(size_t index, size_t size);
 
-	size_t allocate(std::shared_ptr<Process> processAddress) override;
+	int allocate(std::shared_ptr<Process> processAddress) override;
 	void deallocate(std::shared_ptr<Process> processAddress) override;
 
 	size_t getFrameSize() const override { return maxMemorySize; }
